+++ example/SingleSelectionCollectionExample.php
@@ -21,8 +21,16 @@ $collection->append('123','my_key_1');
 echo "Append item 456 using my_key_2 as key\n";
 $collection->append('456','my_key_2');
 
+echo "Check if collection has a selection (must return false)\n";
+
+var_dump($collection->hasSelection());
+
 echo "Select item my_key_1 in collection\n";
 $collection->select('my_key_1');
+
+echo "Check if collection has a selection (must return true)\n";
+
+var_dump($collection->hasSelection());
 
 echo "Is selection locked?\n";
 var_dump($collection->isSelectionLocked());